Collapsed all OpenRTB packages to request and response for go parity
