+++ android/src/main/java/com/adsbynimbus/openrtb/android/AndroidGeoBuilder.java
@@ -5,8 +5,8 @@ import androidx.annotation.IntDef;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
-import com.adsbynimbus.openrtb.targeting.user.Geo;
-import com.adsbynimbus.openrtb.targeting.user.Geo.LocationType;
+import com.adsbynimbus.openrtb.request.Geo;
+import com.adsbynimbus.openrtb.request.Geo.LocationType;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
