+++ android/src/main/java/com/adsbynimbus/openrtb/android/CreativeAttributeInt.java
@@ -2,7 +2,7 @@ package com.adsbynimbus.openrtb.android;
 
 import androidx.annotation.IntDef;
 
-import com.adsbynimbus.openrtb.impression.CreativeAttribute;
+import com.adsbynimbus.openrtb.request.CreativeAttribute;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
