--- java/src/main/java/com/adsbynimbus/openrtb/impression/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Impression OpenRTB models defining the types of ads requested
- */
-package com.adsbynimbus.openrtb.impression;
