--- java/src/main/java/com/adsbynimbus/openrtb/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Top level classes and builder definitions for making a request to Nimbus and receiving response
- */
-package com.adsbynimbus.openrtb;
