+++ java/src/main/java/com/adsbynimbus/openrtb/request/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Provides the classes and values necessary for making an OpenRTB request.
+ *
+ * @since 0.5.0
+ */
+package com.adsbynimbus.openrtb.request;
