+++ java/src/main/java/com/adsbynimbus/openrtb/response/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Provides the classes defining a response to an OpenRTB request.
+ *
+ * @since 0.5.0
+ */
+package com.adsbynimbus.openrtb.response;
