--- java/src/main/java/com/adsbynimbus/openrtb/targeting/distribution/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Required application and publisher models and builders
- */
-package com.adsbynimbus.openrtb.targeting.distribution;
