--- java/src/main/java/com/adsbynimbus/openrtb/targeting/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Targeting information for OpenRTB requests
- */
-package com.adsbynimbus.openrtb.targeting;
