--- java/src/main/java/com/adsbynimbus/openrtb/targeting/user/package-info.java
@@ -1,4 +1,0 @@
-/**
- * User and required Device models and builders
- */
-package com.adsbynimbus.openrtb.targeting.user;
